Edit the RenderIn3D project
Change page format so that it changes the positioning of the canvas and
controls as appropriate to the orientation of the screen.
Fix issues with the Arrows or Vector view so that the arrow are placed along
the 3D grid of points as intended, and that they sample the gradient at their
intended sample point. Begin creating new arrow3d classes that are intended
to replace some of the things in the simulation source file.