More edits to RenderIn3D
Get the arrows grid sliders to properly change the size
dimensionality of the arrows plot.
Account for scrolling in the getNormalizedCanvasCoordinates function.
Fix large gap that occured between the canvas and controls row
that occured on certain browsers when in portrait mode.
Set background to grey when using a browser whose sliders are black.